correction for Parcel and Line model width and length -> double type

+++ greenUDB/Models/Line.cs
@@ -13,7 +13,7 @@ public partial class Line
 
     public long? ParcelId { get; set; }
 
-    public long? Length { get; set; }
+    public double? Length { get; set; }
 
     public DateTime? CreatedAt { get; set; }
